+++ CHANGES.md
@@ -2,6 +2,7 @@
   - Small fix to BibleRangeList.consolidate()
   - Add chap_span() and book_span() to BibleBook, BibleVerse, BibleRange and BibleRangeList.
   - Add BibleRangeList.span()
+  - BibleRange.split() adds regroup bool argument (defaults to True)
   - Remove BibleVerse.chap_count()
   - Documentation fixes
 - v0.10.0: